Update golangci-lint and resolve lint findings. (#114)
* Resolve production lint findings.

Mark existing best-effort cleanup and transaction errors as intentionally ignored without changing their execution order or error-handling policy.

Remove unused parameters and return values, ineffective assignments, redundant conversions, and an unreachable duplicate branch. Correct misspelled source comments.

* Resolve test lint findings.

Remove assignments whose values are overwritten before use. Assert errors from test file writes, and correct misspelled descriptions and variable names.

* Enable additional golangci-lint checks.

Enable ineffassign, misspell, nakedret, SA staticcheck checks except SA1019, and unconvert with read-only module resolution.

Run golangci-lint 2.12.2 through the official GitHub Action. Keep make lint limited to invoking the installed binary and document the workflow.

* Install the pinned golangci-lint version with Go.

Restore make lint bootstrapping without executing a downloaded installation script.

+++ backup/dependencies.go
@@ -239,7 +239,7 @@ func assignCohorts(slice []Sortable, dependencies DependencyMap, isDependentOn m
 	// to the toc.
 	for id := range tierMap {
 		cohort := objectToCohortMap[id]
-		tierMap[id][1] = uint32(cohort)
+		tierMap[id][1] = cohort
 	}
 	return
 }
